// src/char/enums/char.enum.ts
export enum Paths {
  Destruction = 'DESTRUCTION',
  Hunt = 'HUNT',
  Erudition = 'ERUDITION',
  Harmony = 'HARMONY',
  Nihility = 'NIHILITY',
  Preservation = 'PRESERVATION',
  Abundance = 'ABUNDANCE',
}
export enum Types {
  Physical = 'PHYSICAL',
  Fire = 'FIRE',
  Ice = 'ICE',
  Wind = 'WIND',
  Lightning = 'LIGHTNING',
  Quantum = 'QUANTUM',
  Imaginary = 'IMAGINARY',
}